feat: backend api integration for manager register
