+++ src/components/common/LNB/LNB.tsx
@@ -1,8 +1,8 @@
-import { LNBInstructorMenu, LNBManagerMenu } from '@/constants/LNBManage';
+import { LNBInstructorMenu, LNBManagerMenu } from '@/constants';
 import styled from '@emotion/styled';
 import React from 'react';
 import ClaonLogo from '@/assets/ClaonLogo';
-import { LNBMenuDepth1, LNBMenuDepth2 } from '@/types/layout';
+import { LNBMenuDepth1, LNBMenuDepth2 } from '../../../types/layout';
 import { LNBDepth1, LNBDepth2 } from './LNBItems';
 
 const StyledLNBWrapper = styled.div`
