+++ src/hooks/queries/login/queries.ts
@@ -1,4 +1,4 @@
-import { SignInRequest, SignInResponse } from '@/types/login';
+import { SignInRequest, SignInResponse } from '../../../types/login';
 import axios from 'axios';
 import { getSession } from 'next-auth/react';
