+++ src/pages/register/manager/step1.tsx
@@ -21,8 +21,8 @@ import ColorPickerModal from '@/components/register/manager/ColorPickerModal';
 import HoldTypeSelect from '@/components/register/manager/HoldTypeSelect';
 import HoldColorFormItem from '@/components/register/manager/HoldColorFormItem';
 import SearchCenterModal from '@/components/register/manager/SearchCenterModal';
-import { CenterNameResponse } from '@/types/common/center';
 import { useRouter } from 'next/router';
+import { CenterNameResponse } from '../../../types/common/center';
 
 const Title = styled.p`
   font-size: 20px;
